Guest view for LAN + minor view fixes

+++ resources/views/tournament/list_lan.blade.php
@@ -24,12 +24,12 @@
         <td scope="col" class="text-center ">{{ substr(($tournament->desc_tournament), 0, 50) }}</td>
         <td scope="col" class="text-center">{{ $tournament->opening_date_tournament }}</td>
 
- 
+
         <td scope="col" class=" text-center">
             <div class="btn-group">
               <a class="btn btn-success" href="{{ route('tournament.show_tournament', array('lan' => $lan->id, 'tournament' => $tournament->id)) }}"><i class='fa fa-eye'></i></a>
 
-        @if($userIsLanAdmin)
+        @if(isset($userIsLanAdmin) && $userIsLanAdmin)
             <a class="btn btn-warning" href="{{ route('tournament.edit_tournament', array('lan' => $lan->id, 'tournament' => $tournament->id)) }}"><i class='fa fa-edit'></i></a>
 
           {!! Form::open(['method' => 'delete','url'=>'', 'onsubmit'=>'return removeTournament(event,'.$lan->id.','.$tournament->id.')']) !!}
